WIP: Finish dereference-http actor and remove rdf-derference actors
