+++ packages/actor-dereference-http/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/ActorDereferenceHttp';
