--- packages/actor-rdf-dereference-file/lib/index.ts
@@ -1,1 +1,0 @@
-export * from './ActorRdfDereferenceFile';
