--- packages/actor-rdf-dereference-file/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@comunica/actor-rdf-dereference-file",
-  "version": "2.0.0",
-  "description": "A file rdf-dereference actor",
-  "lsd:module": true,
-  "main": "lib/index.js",
-  "typings": "lib/index",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/comunica/comunica.git",
-    "directory": "packages/actor-rdf-dereference-file"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "keywords": [
-    "comunica",
-    "actor",
-    "rdf-dereference",
-    "file"
-  ],
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/comunica/comunica/issues"
-  },
-  "homepage": "https://comunica.dev/",
-  "files": [
-    "components",
-    "lib/**/*.d.ts",
-    "lib/**/*.js"
-  ],
-  "dependencies": {
-    "@comunica/bus-rdf-dereference": "^2.0.0",
-    "@comunica/bus-rdf-parse": "^2.0.0",
-    "@comunica/core": "^2.0.0"
-  },
-  "scripts": {
-    "build": "npm run build:ts && npm run build:components",
-    "build:ts": "node \"../../node_modules/typescript/bin/tsc\"",
-    "build:components": "componentsjs-generator"
-  }
-}
