--- packages/actor-rdf-dereference-file/test/dummy
@@ -1,2 +1,0 @@
-
-:a :b :c.
