--- packages/actor-rdf-dereference-file/test/dummy.ttl
@@ -1,2 +1,0 @@
-
-:a :b :c.
