--- packages/actor-rdf-dereference-file/test/dummy.unknown
@@ -1,2 +1,0 @@
-
-:a :b :c.
