--- packages/actor-rdf-dereference-http/lib/ActorRdfDereferenceHttp.ts
@@ -1,10 +1,0 @@
-import { ActorRdfDereferenceHttpBase } from './ActorRdfDereferenceHttpBase';
-
-/**
- * The non-browser variant of {@link ActorRdfDereferenceHttp}.
- */
-export class ActorRdfDereferenceHttp extends ActorRdfDereferenceHttpBase {
-  protected getMaxAcceptHeaderLength(): number {
-    return this.maxAcceptHeaderLength;
-  }
-}
