--- packages/actor-rdf-dereference-http/lib/index.ts
@@ -1,2 +1,0 @@
-export * from './ActorRdfDereferenceHttpBase';
-export * from './ActorRdfDereferenceHttp';
