--- packages/actor-rdf-dereference-http/package.json
@@ -1,48 +1,0 @@
-{
-  "name": "@comunica/actor-rdf-dereference-http",
-  "version": "2.0.0",
-  "description": "A node-fetch http actor",
-  "lsd:module": true,
-  "main": "lib/index.js",
-  "typings": "lib/index",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/comunica/comunica.git",
-    "directory": "packages/actor-rdf-dereference-http"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "keywords": [
-    "comunica",
-    "runner"
-  ],
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/comunica/comunica/issues"
-  },
-  "homepage": "https://comunica.dev/",
-  "files": [
-    "components",
-    "lib/**/*.d.ts",
-    "lib/**/*.js"
-  ],
-  "dependencies": {
-    "@comunica/bus-http": "^2.0.0",
-    "@comunica/bus-rdf-dereference": "^2.0.0",
-    "@comunica/bus-rdf-parse": "^2.0.0",
-    "@comunica/core": "^2.0.0",
-    "@comunica/mediator-race": "^2.0.0",
-    "cross-fetch": "^3.0.5",
-    "relative-to-absolute-iri": "^1.0.5",
-    "stream-to-string": "^1.2.0"
-  },
-  "scripts": {
-    "build": "npm run build:ts && npm run build:components",
-    "build:ts": "node \"../../node_modules/typescript/bin/tsc\"",
-    "build:components": "componentsjs-generator"
-  },
-  "browser": {
-    "./lib/ActorRdfDereferenceHttp.js": "./lib/ActorRdfDereferenceHttp-browser.js"
-  }
-}
